修复音符条不能触发的bug和按键过窄的情况
1.  修改使得shoot函数发射的时候根据情况把quad的碰撞体组件给失效有效掉
2.  真机测试修改haveTouchOfTheTrack函数的不能触发音符条的bug
3.  分数一致全为float
4.  再次修补音符条一直按着就不会飞的bug（FixedUpdate()函数）
5.修改2160×1080的变色面为z=0.5

+++ Assets/Score/Scripts/Score.cs
@@ -6,7 +6,7 @@ namespace Score
 {
     public static class Score
     {
-        public static double totalScore { get; set; }  //总分
+        public static float totalScore { get; set; }  //总分
 
         public static int missNum { get; set; }   //miss掉的方块数
